Part of WOR-165: step 3 — extract ServiceManager into watcher_services.py
Move _ensure_ollama_running / _ensure_litellm_running / _stop_litellm_proxy
and the _litellm_proc handle into ServiceManager. Backward-compat shims on
Watcher delegate to self._services so existing test patches keep working.
